Move all files out of tests/ folder
