+++ ui/src/AccountDropdown/ViewOnlyAccessFormSection.test.tsx
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-import TestUtils, {renderWithRedux} from '../tests/TestUtils';
+import TestUtils, {renderWithRedux} from '../Utils/TestUtils';
 import React from 'react';
 import {fireEvent, screen, waitFor} from '@testing-library/react';
 import rootReducer from '../Redux/Reducers';
