+++ ui/src/Assignments/AssignmentExistsWarning.test.tsx
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import {renderWithRedux} from '../tests/TestUtils';
+import {renderWithRedux} from '../Utils/TestUtils';
 import AssignmentExistsWarning from './AssignmentExistsWarning';
 import React from 'react';
