+++ ui/src/PeopleMover/ArchivedProduct.test.tsx
@@ -17,8 +17,8 @@
 
 import React from 'react';
 import ArchivedProduct from '../Products/ArchivedProduct';
-import TestUtils, {renderWithRedux} from './TestUtils';
-import PeopleMover from '../PeopleMover/PeopleMover';
+import TestUtils, {renderWithRedux} from '../Utils/TestUtils';
+import PeopleMover from './PeopleMover';
 import {fireEvent, RenderResult, waitFor} from '@testing-library/react';
 import {createBrowserHistory, History} from 'history';
 import {Router} from 'react-router-dom';
