+++ ui/src/Products/Product.test.tsx
@@ -20,7 +20,7 @@ import {act, fireEvent, screen, waitFor} from '@testing-library/react';
 import PeopleMover from '../PeopleMover/PeopleMover';
 import AssignmentClient from '../Assignments/AssignmentClient';
 import ProductClient from './ProductClient';
-import TestUtils, {createDataTestId, renderWithRedux} from '../tests/TestUtils';
+import TestUtils, {createDataTestId, renderWithRedux} from '../Utils/TestUtils';
 import {applyMiddleware, createStore, PreloadedState, Store} from 'redux';
 import rootReducer, {GlobalStateProps} from '../Redux/Reducers';
 import ProductTagClient from '../Tags/ProductTag/ProductTagClient';
