+++ ui/src/TimeOnProductPage/TimeOnProduct.test.tsx
@@ -16,7 +16,7 @@
  */
 
 import React from 'react';
-import TestUtils, {renderWithRedux} from '../tests/TestUtils';
+import TestUtils, {renderWithRedux} from '../Utils/TestUtils';
 import TimeOnProduct, {
     generateTimeOnProductItems,
     LOADING,
